Build SqlStreamStore on Linux
Inspired by the work of Yves from the build-on-linux branch.

- Introduce DockerInstance (Implementation of ILocalInstance)

The `DockerInstance` will be used in place of `LocalInstance` when the platform is detected as `Unix`.

When building the connection string, the `DockerInstance` will favour the environment variables (exposed via docker links).
In the absence of the environment variables, defaults will be used which allows the user to run a local instance of the MsSql docker container for testing purposes.

An example of starting a local MsSql docker container for testing purposes.

```
docker run -d -e 'ACCEPT_EULA=Y' -e 'SA_PASSWORD=SqlStreamSt0re' -p 1433:1433 microsoft/mssql-server-linux:latest
```

- Introduce wercker (wercker.yml, wercker.com)

The wercker script depends on a docker image which houses the dependencies for building SqlStreamStore on linux. Currently the dependencies are `Mono` and `DotNet Core 2.0`.
The docker image (pgermishuys/sqlstreamstoretest) is built via https://github.com/pgermishuys/SQLStreamStoreTest which can be copied to `SqlStreamStore` and the wercker configuration be updated to point to the new docker image.